pyproject.toml
[tool.poetry]
name = "soap_to_rest"
version = "0.1.0"
description = ""
authors = ["Benjamin Brewer <ben@benbrewer.me>"]

[tool.poetry.dependencies]
python = "^3.8"
suds-community = "^0.8.4"
quart = "^0.13.0"
spyne = "^2.13.15"
lxml = "^4.5.2"
schema = "^0.7.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-asyncio = "^0.14.0"
pytest-mockito = "^0.0.4"
pylint = "^2.6.0"
black = "^20.8b1"
isort = "^5.4.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
